talos/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/talos"]

[project]
name = "talos"
version = "0.1.0"
description = "Talos Linux Kubernetes cluster intelligence plugin for EmberAI -- cluster lifecycle, node management, etcd health, diagnostics, and security auditing via talosctl CLI"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
    { name = "Bluminal Labs" },
]
keywords = [
    "talos",
    "talos-linux",
    "kubernetes",
    "k8s",
    "cluster",
    "etcd",
    "talosctl",
    "mcp",
    "emberai",
    "netex",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: System :: Clustering",
    "Topic :: System :: Systems Administration",
    "Typing :: Typed",
    "Framework :: Pydantic :: 2",
]

dependencies = [
    "mcp>=1.26.0,<2",
    "pydantic>=2.12,<3",
    "python-dotenv>=1.2,<2",
    "pyyaml>=6.0,<7",
]

[project.urls]
Homepage = "https://github.com/bluminal/emberai"
Documentation = "https://bluminal.github.io/emberai/talos/"
Repository = "https://github.com/bluminal/emberai/tree/main/talos"
Issues = "https://github.com/bluminal/emberai/issues"
Changelog = "https://github.com/bluminal/emberai/blob/main/docs/reference/changelog.md"

[project.scripts]
talos-server = "talos.server:main"

[project.entry-points."netex.plugins"]
talos = "talos.server:plugin_info"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "coverage>=7.13.5",
    "ruff>=0.15.6",
    "mypy>=1.19.1",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
    "RUF",  # ruff-specific rules
]

[tool.ruff.lint.isort]
known-first-party = ["talos"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/talos"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80
show_missing = true
